requirements.txt
Flask==3.0.3
keras==3.5.0
numpy==1.26.4
matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy==1.26.4
pillow==10.4.0
requests==2.32.3
tensorflow-cpu==2.17.0
Werkzeug==3.0.4

